+++ Plugin.BluetoothLE.Tests/DeviceTests.cs
@@ -26,18 +26,19 @@ namespace Plugin.BluetoothLE.Tests
 
             var count = 0;
             await this.FindTestDevice();
-            this.Device.WhenServiceDiscovered().Subscribe(x =>
-            {
-                count++;
-            });
+            this.Device.WhenServiceDiscovered().Subscribe(_ => count++);
+
             await this.Device.Connect(new GattConnectionConfig
             {
                 AutoConnect = autoConnect,
                 IsPersistent = true
             });
-            await UserDialogs.Instance.AlertAsync("No turn device off - wait a 3 seconds then turn it back on - press OK if light goes green or you believe connection has failed");
-            await this.Device.WhenStatusChanged().FirstAsync(x => x == ConnectionStatus.Connected);
-            //count.Should().Be()
+            await UserDialogs.Instance.AlertAsync("Now turn device off & press OK");
+            var origCount = count;
+            count = 0;
+
+            await UserDialogs.Instance.AlertAsync("No turn device back on & press OK when light turns green");
+            count.Should().Be(origCount);
         }
 
 
@@ -85,28 +86,3 @@ namespace Plugin.BluetoothLE.Tests
         }
     }
 }
-/*
- * this.scan = this.BleAdapter
-    .ScanWhenAdapterReady()
-    //.Where(x => x.AdvertisementData.ServiceUuids.Any(y => y.Equals(ScratchServiceUuid)))
-    .Where(x => x.Device?.Name?.StartsWith("bean", StringComparison.CurrentCultureIgnoreCase) ?? false)
-    .Take(1)
-    .Select(x =>
-    {
-        this.device = x.Device;
-        x.Device.Connect().Subscribe();
-        return x.Device.GetKnownService(ScratchServiceUuid);
-    })
-    .Where(x => x != null)
-    .Switch()
-    .Select(x => x.WhenCharacteristicDiscovered())
-    .Switch()
-    .Subscribe(ch =>
-    {
-        this.WriteMsg("Subscribing to characteristic", ch.Uuid.ToString());
-        ch.RegisterAndNotify().Subscribe(x => this.WriteMsg(
-            x.Characteristic.Uuid.ToString(),
-            UTF8Encoding.UTF8.GetString(x.Data, 0, x.Data.Length)
-        ));
-    });
- */